+++ src/internal/package/getImportMapFromPackageFiles.js
@@ -11,7 +11,7 @@ import { resolvePackageMain } from "./resolvePackageMain.js"
 import { visitPackageExports } from "./visitPackageExports.js"
 import { createFindNodeModulePackage } from "./node-module-resolution.js"
 
-export const getImportMapFromPackages = async ({
+export const getImportMapFromPackageFiles = async ({
   // nothing is actually listening for this cancellationToken for now
   // it's not very important but it would be better to register on it
   // an stops what we are doing if asked to do so
